# frontend/app/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My API",
    "description": "API Documentation with JWT Authentication",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Mailing List",
      "description": "Mailing list signup API"
    }
  ],
  "paths": {
    "/api/profile": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "getProfile",
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/mailing-list/signup": {
      "post": {
        "tags": [
          "Mailing List"
        ],
        "summary": "Sign up for mailing list",
        "description": "Add a new potential customer to the mailing list",
        "operationId": "signupForMailingList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailingListSignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/collectors": {
      "get": {
        "tags": [
          "collector-controller"
        ],
        "operationId": "getAllCollectors",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Collector"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "collector-controller"
        ],
        "operationId": "createCollector",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectorCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Collector"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/artist": {
      "get": {
        "tags": [
          "artist-controller"
        ],
        "operationId": "getAllArtists",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Artist"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "artist-controller"
        ],
        "operationId": "createArtist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/collectors/{id}": {
      "get": {
        "tags": [
          "collector-controller"
        ],
        "operationId": "getCollectorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Collector"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/artist/{id}": {
      "get": {
        "tags": [
          "artist-controller"
        ],
        "operationId": "getArtist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ValidationErrorResponse": {
        "type": "object",
        "description": "Structured response returned when validation fails",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error summary message",
            "example": "Validation failed with 2 errors"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Map of field names to lists of error messages",
            "example": {
              "email": [
                "Invalid format",
                "Email already in use"
              ],
              "password": [
                "Too weak"
              ]
            }
          }
        },
        "required": [
          "errors",
          "message"
        ]
      },
      "ProfileUpdateRequest": {
        "type": "object",
        "description": "User profile information update",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of the user.",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user.",
            "example": "Smith"
          },
          "email": {
            "type": "string",
            "description": "Email of the user.",
            "example": "john.smith@gmail.com"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL to an avatar image for the user.",
            "example": "https://example.com"
          },
          "currentPassword": {
            "type": "string",
            "description": "User\u0027s current password",
            "example": "password123"
          },
          "newPassword": {
            "type": "string",
            "description": "New password for the user",
            "example": "password123"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        },
        "required": [
          "avatarUrl",
          "email",
          "firstName",
          "lastName"
        ]
      },
      "MailingListSignupRequest": {
        "type": "object",
        "description": "Mailing List Signup Request",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "Jane",
            "maxLength": 15,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe",
            "maxLength": 20,
            "minLength": 0
          },
          "areaCode": {
            "type": "string",
            "description": "Area code",
            "example": 864,
            "maxLength": 3,
            "minLength": 0
          },
          "telephoneNumber": {
            "type": "string",
            "description": "Telephone number",
            "example": 5551234,
            "maxLength": 7,
            "minLength": 0
          },
          "street": {
            "type": "string",
            "description": "Street address",
            "example": 123,
            "maxLength": 50,
            "minLength": 0
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "Greenville"
          },
          "state": {
            "type": "string",
            "description": "State",
            "example": "SC"
          },
          "zipCode": {
            "type": "string",
            "description": "Zip code",
            "example": 29601
          },
          "preferredArtistFirstName": {
            "type": "string",
            "description": "Preferred artist first name",
            "example": "Pablo"
          },
          "preferredArtistLastName": {
            "type": "string",
            "description": "Preferred artist last name",
            "example": "Picasso"
          },
          "preferredMedium": {
            "type": "string",
            "description": "Preferred medium",
            "example": "Watercolor",
            "maxLength": 15,
            "minLength": 0
          },
          "preferredStyle": {
            "type": "string",
            "description": "Preferred style",
            "example": "Impressionist",
            "maxLength": 15,
            "minLength": 0
          },
          "preferredType": {
            "type": "string",
            "description": "Preferred type",
            "example": "Landscape",
            "maxLength": 20,
            "minLength": 0
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "CollectorCreateRequest": {
        "type": "object",
        "description": "Collector Creation Request",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "Bob",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Smith",
            "minLength": 1
          },
          "interviewDate": {
            "type": "string",
            "format": "date",
            "description": "Date of interview",
            "example": "2023-01-01"
          },
          "interviewerName": {
            "type": "string",
            "description": "Interviewer name",
            "example": "Kevin Treu"
          },
          "areaCode": {
            "type": "string",
            "description": "Area code",
            "example": 123
          },
          "telephoneNumber": {
            "type": "string",
            "description": "Telephone number",
            "example": 1234567
          },
          "street": {
            "type": "string",
            "description": "Street address",
            "example": 123
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "Anytown"
          },
          "state": {
            "type": "string",
            "description": "State",
            "example": "CA"
          },
          "zipCode": {
            "type": "string",
            "description": "Zip code",
            "example": 12345
          },
          "socialSecurityNumber": {
            "type": "string",
            "description": "Social security number",
            "example": "123-45-6789"
          },
          "preferredArtistFirstName": {
            "type": "string",
            "description": "Preferred artist first name",
            "example": "Pablo"
          },
          "preferredArtistLastName": {
            "type": "string",
            "description": "Preferred artist last name",
            "example": "Picasso"
          },
          "collectionMedium": {
            "type": "string",
            "description": "Collection medium",
            "example": "Oil"
          },
          "collectionStyle": {
            "type": "string",
            "description": "Collection style",
            "example": "Abstract"
          },
          "collectionType": {
            "type": "string",
            "description": "Collection type",
            "example": "Painting"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "socialSecurityNumber"
        ]
      },
      "Artist": {
        "type": "object",
        "properties": {
          "artistId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "interviewDate": {
            "type": "string",
            "format": "date"
          },
          "interviewerName": {
            "type": "string"
          },
          "areaCode": {
            "type": "string"
          },
          "telephoneNumber": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "zip": {
            "$ref": "#/components/schemas/Zip"
          },
          "salesLastYear": {
            "type": "number"
          },
          "salesYearToDate": {
            "type": "number"
          },
          "socialSecurityNumber": {
            "type": "string"
          },
          "usualMedium": {
            "type": "string"
          },
          "usualStyle": {
            "type": "string"
          },
          "usualType": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "Collector": {
        "type": "object",
        "properties": {
          "socialSecurityNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "interviewDate": {
            "type": "string",
            "format": "date"
          },
          "interviewerName": {
            "type": "string"
          },
          "areaCode": {
            "type": "string"
          },
          "telephoneNumber": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "zip": {
            "$ref": "#/components/schemas/Zip"
          },
          "salesLastYear": {
            "type": "number"
          },
          "salesYearToDate": {
            "type": "number"
          },
          "collectionArtistId": {
            "$ref": "#/components/schemas/Artist"
          },
          "collectionMedium": {
            "type": "string"
          },
          "collectionStyle": {
            "type": "string"
          },
          "collectionType": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "socialSecurityNumber"
        ]
      },
      "Zip": {
        "type": "object",
        "properties": {
          "zip": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "city",
          "state",
          "zip"
        ]
      },
      "AuthRequest": {
        "type": "object",
        "description": "API Login and Authorization Request",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username",
            "example": "admin",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "admin123",
            "maxLength": 2147483647,
            "minLength": 6
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT Authorization Token",
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOiJhYmNkMTIzIiwiZXhwaXJ5IjoxNjQ2NjM1NjExMzAxfQ.3Thp81rDFrKXr3WrY1MyMnNK8kKoZBX9lg-JwFznR-M"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "admin"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "admin@furman.edu"
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "token",
          "username"
        ]
      },
      "ArtistCreateRequest": {
        "type": "object",
        "description": "Artist Creation Request",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "Bob",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Smith",
            "minLength": 1
          },
          "interviewDate": {
            "type": "string",
            "format": "date",
            "description": "Date of interview",
            "example": "2023-01-01"
          },
          "interviewerName": {
            "type": "string",
            "description": "Interviewer name",
            "example": "Kevin Treu"
          },
          "areaCode": {
            "type": "string",
            "description": "Area code",
            "example": 123
          },
          "telephoneNumber": {
            "type": "string",
            "description": "Telephone number",
            "example": 1234567
          },
          "street": {
            "type": "string",
            "description": "Street address",
            "example": 123
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "Anytown"
          },
          "state": {
            "type": "string",
            "description": "State",
            "example": "CA"
          },
          "zipCode": {
            "type": "string",
            "description": "Zip code",
            "example": 12345
          },
          "socialSecurityNumber": {
            "type": "string",
            "description": "Social security number",
            "example": "123-45-6789"
          },
          "usualMedium": {
            "type": "string",
            "description": "Usual medium",
            "example": "Oil"
          },
          "usualStyle": {
            "type": "string",
            "description": "Usual style",
            "example": "Abstract"
          },
          "usualType": {
            "type": "string",
            "description": "Usual type",
            "example": "Painting"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Enter JWT token in the format: Bearer {token}",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}